chore: move index.js to top level dir
