+++ examples/schema_test.js
@@ -1,9 +1,8 @@
-import { encode, decode } from '../src/index.js'
+import { encode, decode } from '../index.js'
 import Hypercore from 'hypercore'
 import ram from 'random-access-memory'
 import { randomBytes } from 'node:crypto'
-
-const record = encode({
+const obj = {
   id: randomBytes(32).toString('hex'),
   type: 'observation',
   schemaVersion: 4,
@@ -14,7 +13,8 @@ const record = encode({
   metadata: {
     manual_location: true,
   },
-})
+}
+const record = encode(obj)
 
 const core = new Hypercore(ram, { valueEncoding: 'binary' })
 await core.ready()
@@ -23,7 +23,6 @@ core.append(record)
 try {
   const index = 0
   const data = await core.get(index)
-  console.log(decode(data, { key: core.key, index }))
   if (Buffer.compare(data, record) !== 0) {
     throw new Error(`data doesn't match: ${data} != ${record}`)
   } else {
